backend/requirements.txt
fastapi>=0.116.1
uvicorn>=0.35.0
supabase>=2.3.4
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.0,<5.0.0
python-multipart>=0.0.6
pydantic>=2.0.0
pydantic-settings>=2.0.0
httpx>=0.27.0
python-dotenv>=1.0.0
openai>=1.12.0
